+++ glean-core/android/src/main/java/mozilla/telemetry/glean/debug/GleanDebugActivity.kt
@@ -5,6 +5,7 @@
 package mozilla.telemetry.glean.debug
 
 import android.app.Activity
+import android.content.ComponentName
 import android.content.Intent
 import android.os.Bundle
 import android.util.Log
@@ -37,7 +38,20 @@ class GleanDebugActivity : Activity() {
          * Tags all outgoing pings as debug pings to make them available for real-time validation.
          * The value must match the pattern `[a-zA-Z0-9-]{1,20}`.
          */
-        const val TAG_DEBUG_VIEW_EXTRA_KEY = "tagPings"
+        const val TAG_DEBUG_VIEW_EXTRA_KEY = "debugViewTag"
+        const val LEGACY_TAG_PINGS = "tagPings"
+
+        /**
+         * Tags all outgoing pings as debug pings to make them available for real-time validation.
+         * The value must match the pattern `[a-zA-Z0-9-]{1,20}`.
+         */
+        const val SOURCE_TAGS_KEY = "sourceTags"
+
+        /**
+         * Executes the activity with the provided name instead of the main one
+         * after finishing with the `GleanDebugActivity`.
+         */
+        const val NEXT_ACTIVITY_TO_RUN = "startNext"
     }
 
     // IMPORTANT: These activities are unsecured, and may be triggered by
@@ -49,6 +63,7 @@ class GleanDebugActivity : Activity() {
     /**
      * On creation of the debug activity, launch the requested command.
      */
+    @Suppress("ComplexMethod", "LongMethod")
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
 
@@ -68,8 +83,12 @@ class GleanDebugActivity : Activity() {
         }
 
         // Make sure that at least one of the supported commands was used.
-        val supportedCommands =
-            listOf(SEND_PING_EXTRA_KEY, LOG_PINGS_EXTRA_KEY, TAG_DEBUG_VIEW_EXTRA_KEY)
+        val supportedCommands = listOf(
+            SEND_PING_EXTRA_KEY, LOG_PINGS_EXTRA_KEY, NEXT_ACTIVITY_TO_RUN,
+            TAG_DEBUG_VIEW_EXTRA_KEY, SOURCE_TAGS_KEY, LEGACY_TAG_PINGS
+        )
+
+        var nextActivityName: String? = null
 
         // Enable debugging options and start the application.
         intent.extras?.let {
@@ -86,6 +105,13 @@ class GleanDebugActivity : Activity() {
             // Set the debug view tag, if the tag is invalid it won't be set
             debugViewTag?.let {
                 Glean.setDebugViewTag(debugViewTag)
+            } ?: run {
+                // If the 'debugViewTag' was not used, try to look for the legacy
+                // way of setting debug view tags. We should leave this block of
+                // code around at most until December 2020.
+                intent.getStringExtra(LEGACY_TAG_PINGS)?.let { legacyTag ->
+                    Glean.setDebugViewTag(legacyTag)
+                }
             }
 
             val logPings: Boolean? = intent.getBooleanExtra(LOG_PINGS_EXTRA_KEY, false)
@@ -93,22 +119,42 @@ class GleanDebugActivity : Activity() {
                 Glean.setLogPings(logPings)
             }
 
-            intent.getStringExtra(SEND_PING_EXTRA_KEY)?.let {
-                Glean.submitPingByName(it)
+            intent.getStringArrayExtra(SOURCE_TAGS_KEY)?.let { tags ->
+                Glean.setSourceTags(tags.toSet())
+            }
+
+            intent.getStringExtra(NEXT_ACTIVITY_TO_RUN)?.let { nextActivity ->
+                nextActivityName = nextActivity
+            }
+
+            // Important: this should be applied as the last one, so that
+            // any other option will affect the ping submission as well.
+            intent.getStringExtra(SEND_PING_EXTRA_KEY)?.let { name ->
+                Glean.submitPingByName(name)
             }
         }
 
         // This Activity can be used to tag tests on CI or start products with specific
         // options. We need to make sure to retain and propagate all the options that
-        // we were passed to the next intent. Our strategy:
-        // - use the `Intent` copy constructor to copy all the intent options from the
+        // we were passed to the next intent. Our 3 steps strategy:
+        // 1. use the `Intent` copy constructor to copy all the intent options from the
         //   intent which started this activity;
-        // - get the main launch intent for the product using the Glean SDK;
-        // - change the starting "component" and package to the one from the previous step.
         val nextIntent = Intent(intent)
-        val defaultLaunchIntent = packageManager.getLaunchIntentForPackage(packageName)!!
-        nextIntent.component = defaultLaunchIntent.component
-        nextIntent.`package` = defaultLaunchIntent.`package`
+
+        // 2. get the main launch intent for the product using the Glean SDK or lookup
+        //    the one passed as a command line argument to this activity.
+        val launchIntent = packageManager.getLaunchIntentForPackage(packageName)!!
+        nextIntent.`package` = launchIntent.`package`
+
+        val nextComponent = nextActivityName?.let { name ->
+            // Try to lookup the Activity that was asked by the caller.
+            ComponentName(packageName, name)
+        } ?: launchIntent.component
+
+        Log.i(LOG_TAG, "Running next: ${nextComponent!!.packageName}/${nextComponent.className}")
+
+        // 3. change the starting "component" to the one from the previous step.
+        nextIntent.component = nextComponent
         startActivity(nextIntent)
 
         finish()